Splits out ObscuroScan API methods.
